fix(ui): schedule limit loophole refactored
