+++ src/views/components/PopupMain.tsx
@@ -1,5 +1,4 @@
 import splashText from '@assets/insideJokes';
-import createSchedule from '@pages/background/lib/createSchedule';
 import { background } from '@shared/messages';
 import { initSettings, OptionsStore } from '@shared/storage/OptionsStore';
 import { UserScheduleStore } from '@shared/storage/UserScheduleStore';
@@ -19,7 +18,7 @@ import SettingsIcon from '~icons/material-symbols/settings';
 
 import { Button } from './common/Button';
 import CourseStatus from './common/CourseStatus';
-import { usePrompt } from './common/DialogProvider/DialogProvider';
+import { useAddSchedule } from './common/DialogProvider/AddSchedule';
 import { SmallLogo } from './common/LogoIcon';
 import PopupCourseBlock from './common/PopupCourseBlock';
 import ScheduleDropdown from './common/ScheduleDropdown';
@@ -62,7 +61,8 @@ export default function PopupMain(): JSX.Element {
     const [activeSchedule, schedules] = useSchedules();
     const [isRefreshing, setIsRefreshing] = useState(false);
     const [funny, setFunny] = useState<string>('');
-    const showDialog = usePrompt();
+
+    const handleAddSchedule = useAddSchedule();
 
     useEffect(() => {
         const randomIndex = Math.floor(Math.random() * splashText.length);
@@ -79,32 +79,6 @@ export default function PopupMain(): JSX.Element {
     const handleCalendarOpenOnClick = async () => {
         await background.switchToCalendarTab({});
         window.close();
-    };
-
-    const handleAddSchedule = () => {
-        if (schedules.length >= 10) {
-            showDialog({
-                title: `You have 10 active schedules!`,
-
-                description: (
-                    <>
-                        To encourage organization,{' '}
-                        <span className='text-ut-burntorange'>please consider removing some unused schedules</span> you
-                        may have.
-                    </>
-                ),
-                // eslint-disable-next-line react/no-unstable-nested-components
-                buttons: close => (
-                    <Button variant='filled' color='ut-burntorange' onClick={close}>
-                        I Understand
-                    </Button>
-                ),
-            });
-
-            return;
-        }
-
-        createSchedule('New Schedule');
     };
 
     return (
