+++ src/views/components/calendar/CalendarSchedules.tsx
@@ -1,4 +1,3 @@
-import createSchedule from '@pages/background/lib/createSchedule';
 import { UserScheduleStore } from '@shared/storage/UserScheduleStore';
 import { Button } from '@views/components/common/Button';
 import List from '@views/components/common/List';
@@ -9,7 +8,7 @@ import React from 'react';
 
 import AddSchedule from '~icons/material-symbols/add';
 
-import { usePrompt } from '../common/DialogProvider/DialogProvider';
+import { useAddSchedule } from '../common/DialogProvider/AddSchedule';
 
 /**
  * Renders a component that displays a list of schedules.
@@ -19,33 +18,8 @@ import { usePrompt } from '../common/DialogProvider/DialogProvider';
  */
 export function CalendarSchedules() {
     const [, schedules] = useSchedules();
-    const showDialog = usePrompt();
 
-    const handleAddSchedule = () => {
-        if (schedules.length >= 10) {
-            showDialog({
-                title: `You have 10 active schedules!`,
-
-                description: (
-                    <>
-                        To encourage organization,{' '}
-                        <span className='text-ut-burntorange'>please consider removing some unused schedules</span> you
-                        may have.
-                    </>
-                ),
-                // eslint-disable-next-line react/no-unstable-nested-components
-                buttons: close => (
-                    <Button variant='filled' color='ut-burntorange' onClick={close}>
-                        I Understand
-                    </Button>
-                ),
-            });
-
-            return;
-        }
-
-        createSchedule('New Schedule');
-    };
+    const handleAddSchedule = useAddSchedule();
 
     return (
         <div className='min-w-full w-0 items-center'>
@@ -53,7 +27,12 @@ export function CalendarSchedules() {
                 <Text variant='h3' className='text-nowrap'>
                     MY SCHEDULES
                 </Text>
-                <Button variant='single' color='theme-black' className='h-fit p-0 btn' onClick={handleAddSchedule}>
+                <Button
+                    variant='single'
+                    color='theme-black'
+                    className='h-fit p-0 btn'
+                    onClick={() => handleAddSchedule()}
+                >
                     <AddSchedule className='h-6 w-6' />
                 </Button>
             </div>
